Корректирует названия файлов
