+++ js/views/view-modal-confirm.js
