+++ js/views/view-question-classify.js
@@ -4,7 +4,7 @@ import statsBarTemplate from '../templates/template-statsBar';
 // logic
 import resizeImage from '../data/resize-image';
 
-export default class QuestionViewRadio extends AbstractView {
+export default class QuestionViewClassify extends AbstractView {
   constructor(question, gameState) {
     super();
     this.question = question;
@@ -15,11 +15,20 @@ export default class QuestionViewRadio extends AbstractView {
     return `
     <div class="game">
     <p class="game__task">${this.question.description}</p>
-    <form class="game__content game__content--triple">
+    <form class="${this.question.inner}">
     ${[...this.question.params].map((param) => `
-      <div class="game__option" data-type="${param.class}">
-      <img src="${param.src}" alt="Option 1" width="304" height="455">
-      </div>`).join(``)}
+      <div class="game__option" data-type="${param.class}" data-number="${param.index}">
+      <img src="${param.src}" alt="Option ${param.index}" width="705" height="455">
+      <label class="game__answer  game__answer--photo">
+      <input name="question${param.index}" type="radio" value="photo">
+      <span>Фото</span>
+      </label>
+      <label class="game__answer  game__answer--wide  game__answer--paint">
+      <input name="question${param.index}" type="radio" value="paint">
+      <span>Рисунок</span>
+      </label>
+      </div>
+      `).join(``)}
     </form>
     ${statsBarTemplate(this.gameState)}
     </div>
@@ -52,22 +61,35 @@ export default class QuestionViewRadio extends AbstractView {
     const images = this.element.querySelectorAll(`.game__option > img`);
     images.forEach((image) => {
       image.parentNode.style.display = `none`;
-      image.style.pointerEvents = `none`; // для firefox click
 
       image.addEventListener(`load`, () => {
         this.onGameImageLoad(image);
       });
     });
 
+    const form = this.element.querySelector(`.game__content`);
     const options = this.element.querySelectorAll(`.game__option`);
 
-    options.forEach((option) => {
-      option.addEventListener(`click`, (evt) => {
-        const target = evt.target;
-        const result = target.dataset.type === this.question.answerCorrect;
+    form.addEventListener(`change`, () => {
 
-        this.onAnswer(result);
+      let answers = [];
+
+      options.forEach((option) => {
+        let optionValue = option.dataset.type;
+        let versions = option.querySelectorAll(`input`);
+
+        versions.forEach((version) => {
+          if (version.checked) {
+            let answerValue = version.value;
+            answers.push(optionValue === answerValue);
+          }
+        });
       });
+
+      if (options.length === answers.length) {
+        const result = !answers.includes(false);
+        this.onAnswer(result);
+      }
     });
   }
 }
